backend/worker/requirements.txt
firebase-admin>=6.2.0
google-cloud-firestore>=2.21.0
google-cloud-storage>=2.10.0
rembg>=2.0.67
onnxruntime>=1.22.0
Pillow>=10.0.1
requests>=2.32.5
PyMatting>=1.1.14

